pyproject.toml
[tool.poetry]
name = "capacities-python"
version = "0.1.0"
description = "Collection of tools to interact with your Capacities spaces via the API."
authors = ["Jorge Gómez Zarzosa <jorge.gomzar@gmail.com>"]
readme = "README.md"
packages = [
  {include = "capacities"}
]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
python-telegram-bot = "^21.6"
pydantic = "^2.9.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.382.post1"
ruff-lsp = "^0.0.57"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
